refactor : atomic UI 폴더 이동
