+++ src/pages/IssueDetailPage.tsx
@@ -2,12 +2,11 @@ import ReactMarkdown from 'react-markdown';
 import { useParams } from 'react-router-dom';
 
 import { getIssue } from '@/apis';
-import { Avatar } from '@/components/Avatar';
 import { IssueItem } from '@/components/IssueItem';
-import { LoadingSpinner } from '@/components/LoadingSpinner';
+import { Avatar } from '@/components/ui/Avatar';
+import { LoadingSpinner } from '@/components/ui/LoadingSpinner';
 import { useFetch } from '@/hooks/useFetch.ts';
 import * as S from './IssueDetailPage.styles.tsx';
-import { IssueDetailContentWrapper } from './IssueDetailPage.styles.tsx';
 
 export function IssueDetailPage() {
   const params = useParams<{ id: string }>();
@@ -24,7 +23,7 @@ export function IssueDetailPage() {
         <Avatar src={user?.avatar_url || ''} />
         <IssueItem issue={issue} />
       </S.IssueDetailHeader>
-      <IssueDetailContentWrapper>{body && <ReactMarkdown children={body} />}</IssueDetailContentWrapper>
+      <S.IssueDetailContentWrapper>{body && <ReactMarkdown children={body} />}</S.IssueDetailContentWrapper>
     </main>
   );
 }
